move DummyQuizQuestion into dungeon/test
