+++ build.gradle
@@ -51,7 +51,7 @@ sourceCompatibility = 17
 [compileJava, compileTestJava]*.options*.encoding = "UTF-8"
 sourceSets.main.java.srcDirs = ["game/src/", "dsl/src/", "dungeon/src", "$projectDir/build/generated-src/"]
 sourceSets.main.resources.srcDirs = ["game/assets/"]
-sourceSets.test.resources.srcDirs = ["dsl/test_resources"]
+sourceSets.test.resources.srcDirs = ["dsl/test_resources", "dungeon/assets/"]
 sourceSets.test.java.srcDirs = ["game/test/", "dsl/test/", "dungeon/test"]
 sourceSets.main.antlr.srcDirs = ["dsl/src/antlr"]
