+++ dungeon/test/assets/image_quiz/dummy.png
