+++ game/src/core/utils/Constants.java
@@ -47,7 +47,7 @@ public final class Constants {
 
     public static final String QUIZ_MESSAGE_SOLUTION = "Lösung";
 
-    public static final String TEST_IMAGE_PATH_FOR_DIALOG = "image_quiz/dummy.png";
+    public static final String TEST_IMAGE_PATH_FOR_DIALOG = "dungeon/assets/image_quiz/dummy.png";
 
     /**
      * @param path the relative path to the resource
